+++ app/models/external_income.rb
@@ -1,0 +1,9 @@
+class ExternalIncome < ApplicationRecord
+  include PublicActivity::Common
+  include HasFinancialQuarter
+
+  belongs_to :organisation
+  belongs_to :activity
+
+  validates_presence_of :organisation_id, :financial_quarter, :financial_year, :amount
+end